cr fixes, refactoring
