+++ src/components/index.ts
@@ -23,6 +23,6 @@ export * from './Dialogs'
 export { default as StudioContainer } from './StudioContainer'
 export * from './Topbar'
 export { default as NoConnectionIndicator } from './NoConnectionIndicator'
-export * from './SignInSteps'
+export { default as SignInStepsStepper } from './SignInSteps'
 export * from './StudioEntrypoint'
 export * from './PrivateRoute'
